requirements.txt
deepface==0.0.93
mss==10.0.0
numpy==2.2.6
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
playsound==1.3.0
protobuf==6.31.1
PyAudio==0.2.14
python-dotenv==1.1.0
torch==2.7.0
ultralytics==8.3.126
google-genai
googleapis-common-protos
tf-keras
PySide6